Use a heap
The previous solution relied on std::partial_sort to get the max n
elements. The disadvantage to this is that I am using linear storage,
and the partial_sort operation is therefore logarithmic in complexity.

If a (max) heap is used instead, then for N = 3, constant complexity and
storage is achieved. A quirk of asking for the max 3 values is used in
the implementation: there is only ever 4 elements in the (max) heap at
any one time. This means that selecting the minimum element, moving it
to the end of the vector, and then calling std::vector::pop_back still
preserves the heap property. This is not true for all values of N.